+++ CHANGES.rst
@@ -27,6 +27,7 @@ Improvements
 - Log changes to event reminders in the event log (:pr:`7242`)
 - Allow sending account creation notifications to specific email addresses (:issue:`7166`,
   :pr:`7233`, thanks :user:`duartegalvao`)
+- Add a 'reviewer proposals' column to the abstract list (:pr:`7258`)
 
 Bugfixes
 ^^^^^^^^